+++ src/__tests__/Search.test.tsx
@@ -367,7 +367,7 @@ describe("Renders Search Results Page", () => {
       test("Shows license with links", () => {
         expect(
           screen.getByText("License: Public Domain").closest("a").href
-        ).toContain("/license");
+        ).toContain("/copyright");
       });
       test("Shows cover", () => {
         expect(
@@ -416,7 +416,7 @@ describe("Renders Search Results Page", () => {
       test("Shows Unknown license with links", () => {
         expect(
           screen.getByText("License: Unknown").closest("a").href
-        ).toContain("/license");
+        ).toContain("/copyright");
       });
       test("Shows Placeholder cover", () => {
         expect(
@@ -486,7 +486,7 @@ describe("Renders Search Results Page", () => {
               "License: Public Domain Public Domain Public Domain Public Domain Public Domain Public Domain Public Domain Public Domain Public Domain"
             )
             .closest("a").href
-        ).toContain("/license");
+        ).toContain("/copyright");
       });
       test("Shows cover", () => {
         expect(